chore: rework decorators directory
